// frontend/src/enums/sort-field.enum.ts
export enum SortField {
    ID = "_id",
    NAME = "name",
    SURNAME = "surname",
    EMAIL = "email",
    AGE = "age",
    COURSE = "course",
    FORMAT = "format",
    TARIFF = "tariff",
    GROUP_ID = "groupId",
    PHONE = "phone",
    STATUS = "status",
    ALREADY_PAID = "alreadyPaid",
    SUM = "sum",
    CREATED_AT = "createdAt",
    MANAGER_ID = "managerId",
}

export enum OrderDirection {
    ASC = 'asc',
    DESC = 'desc',
}